+++ opentelemetry-zipkin/src/exporter/mod.rs
@@ -97,8 +97,16 @@ impl ZipkinExporterBuilder {
     /// When using this method, the export timeout will depend on the provided
     /// client implementation and may not respect the timeout set via the
     /// environment variable `OTEL_EXPORTER_ZIPKIN_TIMEOUT`.
-    pub fn with_http_client<T: HttpClient + 'static>(mut self, client: T) -> Self {
-        self.client = Some(Arc::new(client));
+    pub fn with_http_client<T: HttpClient + 'static>(self, client: T) -> Self {
+        self.with_shared_http_client(client.into())
+    }
+
+    /// Assign client implementation wrapped into shared pointer
+    pub fn with_shared_http_client<T: HttpClient + 'static>(
+        mut self,
+        client: std::sync::Arc<T>,
+    ) -> Self {
+        self.client = Some(client);
         self
     }